requirements.txt
streamlit
pandas
matplotlib
sklearn.linear_model
numpy